Separate Junction classes
