+++ src/Test/Junction.cpp
@@ -4,11 +4,15 @@ bool Junction::connectsTo(Junction const& dest, Direction &dir) {
   for (int i = 0; i < 4; i++) {
     dir = (Direction)i;
     Junction const* j = next(dir);
-    if (j != NULL && j->_id == dest._id) {
+    if (j && (j->_id == dest._id)) {
       return true;
     }
   }
   return false;
+}
+
+JunctionId Junction::id() const {
+  return _id;
 }
 
 Junction const* Junction::next(Direction dir) const {
@@ -37,34 +41,4 @@ void Junction::Link(Junction &prev, Junction &next, Direction dir, bool twoway) 
     prev._prevs[opp] = &next;
     next._nexts[opp] = &prev;
   }
-}
-
-/* Definitions for static class Junctions */
-
-Junction Junctions::A;
-Junction Junctions::B;
-Junction Junctions::C;
-Junction Junctions::D;
-Junction Junctions::E;
-Junction Junctions::F;
-Junction Junctions::G;
-Junction Junctions::H;
-
-void Junctions::Init() {
-  Junction::Link(A, B, East, true);
-  // Junction::Link(B, C, East, true); // Not allowed in either direction
-  Junction::Link(C, D, East, true);
-      
-  Junction::Link(A, E, South, true);
-  Junction::Link(B, F, South, true);
-  Junction::Link(C, G, South, true);
-  Junction::Link(D, H, South, true);
-
-  Junction::Link(E, F, East, true);
-  Junction::Link(F, G, East, true);
-  Junction::Link(G, H, East, true);
-}
-
-Junction const *const Junctions::ById(JunctionId id) {
-  return All[id];
 }
